design: 옵션 모달 퍼블리싱 완료
