+++ frontend/src/components/NewChecklist/OptionModal/OptionModal.tsx
@@ -1,20 +1,18 @@
 import styled from '@emotion/styled';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import Button from '@/components/common/Button/Button';
-import Checkbox from '@/components/common/Checkbox/Checkbox';
 import Modal from '@/components/common/Modal/Modal';
 import ModalFooter from '@/components/common/Modal/ModalFooter';
 import OptionButton from '@/components/common/OptionButton/OptionButton';
-import { flexCenter, flexColumn, title4 } from '@/styles/common';
-import theme from '@/styles/theme';
+import OptionModalInfoBox from '@/components/NewChecklist/OptionModal/OptionModalInfoBox';
+import { flexColumn } from '@/styles/common';
 
-const totalOptionCount = 14;
+export const totalOptionCount = 14;
 
 const OptionModal = () => {
   const [isOpen, setIsOpen] = useState(true);
   const [selectedOptions, setSelectedOptions] = useState<number[]>([]);
-  const [isAllSelected, setIsAllSelected] = useState(selectedOptions.length === totalOptionCount);
 
   const onCloseModal = () => {
     setIsOpen(false);
@@ -33,19 +31,11 @@ const OptionModal = () => {
     }
   };
 
-  const onClickSelectAllOptions = () => {
-    if (isAllSelected) {
-      return setSelectedOptions([]);
-    }
-    const allOptions = new Array(totalOptionCount).fill(0).map((e, i) => i + 1);
-    return setSelectedOptions(allOptions);
-  };
-
   const optionCounts = new Array(totalOptionCount).fill(0).map((e, i) => i + 1);
 
-  useEffect(() => {
-    setIsAllSelected(selectedOptions.length === totalOptionCount);
-  }, [selectedOptions]);
+  // useEffect(() => {
+  //   setIsAllSelected(selectedOptions.length === totalOptionCount);
+  // }, [selectedOptions]);
 
   return (
     <div>
@@ -53,18 +43,8 @@ const OptionModal = () => {
         <Modal.header title={'방에 포함된 옵션을 선택해주세요.'}></Modal.header>
         <Modal.body>
           <S.FlexBoxColumn>
-            <S.ButtonContainer>
-              <S.TotalSelectBox>
-                <Checkbox
-                  isChecked={isAllSelected}
-                  onClick={onClickSelectAllOptions}
-                  color={theme.palette.yellow500}
-                  hoverBorderColor={theme.palette.yellow600}
-                />
-                <span>전체선택</span>
-              </S.TotalSelectBox>
-            </S.ButtonContainer>
-            <S.Container>
+            <OptionModalInfoBox selectedOptions={selectedOptions} setSelectedOptions={setSelectedOptions} />
+            <S.OptionContainer>
               {optionCounts.map(optionId => (
                 <OptionButton
                   onClickSelect={targetId => onToggleSelectOption(targetId)}
@@ -73,7 +53,7 @@ const OptionModal = () => {
                   isSelected={isSelectedOption(optionId)}
                 />
               ))}
-            </S.Container>
+            </S.OptionContainer>
           </S.FlexBoxColumn>
         </Modal.body>
         <ModalFooter>
@@ -86,39 +66,24 @@ const OptionModal = () => {
 
 export default OptionModal;
 
-const Container = styled.div`
+const OptionContainer = styled.div`
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   gap: 15px;
-  width: 100%;
+  width: 300px;
+
   justify-content: center;
   place-items: center;
 `;
 
-const ButtonContainer = styled.div`
-  display: flex;
-  width: 100%;
-  height: 50px;
-  justify-content: end;
-`;
-
 const FlexBoxColumn = styled.div`
   ${flexColumn}
-`;
+  width:100%;
 
-const TotalSelectBox = styled.div`
-  ${flexCenter}
-  gap:10px;
-
-  span {
-    ${title4};
-    color: ${({ theme }) => theme.palette.grey600};
-  }
+  align-items: center;
 `;
 
 const S = {
-  Container,
-  TotalSelectBox,
+  OptionContainer,
   FlexBoxColumn,
-  ButtonContainer,
 };
