+++ frontend/src/components/common/Modal/ModalBody.tsx
@@ -16,6 +16,7 @@ export type ModalContentPosition = 'center' | 'left';
 const Container = styled.div<{ contentPosition: ModalContentPosition }>`
   display: flex;
   margin-bottom: 10px;
+  padding: 0 16px;
 
   line-height: 1.5;
   text-align: left;
