+++ frontend/src/components/common/SelectionCounter/SelectionCounter.tsx
@@ -1,5 +1,6 @@
 import styled from '@emotion/styled';
 
+import { title4 } from '@/styles/common';
 import theme from '@/styles/theme';
 
 interface Props {
@@ -33,6 +34,7 @@ const Container = styled.div<{ $color: string; $gap: number }>`
 
   color: ${({ $color }) => $color};
   gap: ${({ $gap }) => $gap}px;
+  ${title4}
 `;
 
 const S = {
