+++ frontend/src/pages/ModalPage.tsx
@@ -1,4 +1,4 @@
-import OptionModal from '@/components/NewChecklist/OptionModal';
+import OptionModal from '@/components/NewChecklist/OptionModal/OptionModal';
 
 export const ModalPage = () => {
   return <OptionModal />;
